+++ Helltooth/src/ht/graphics/rendering/renderers/2D/BatchRenderer2D.cpp
@@ -22,11 +22,11 @@ namespace ht { namespace graphics {
 		buffer = (VertexData*)glMapBuffer(GL_ARRAY_BUFFER, GL_WRITE_ONLY);
 	}
 
-	unsigned int BatchRenderer2D::submitTexture(float id) {
-		unsigned int tid = 0;
-		if (id  > 0.5) {
+	uint32 BatchRenderer2D::submitTexture(f32 id) {
+		uint32 tid = 0;
+		if (id > 0.5f) {
 			bool found = false;
-			for (unsigned int i = 0; i < tids.size(); i++)
+			for (uint32 i = 0; i < tids.size(); i++)
 				if (tids[i] == id) {
 					found = true;
 					tid = i + 1;
@@ -46,9 +46,9 @@ namespace ht { namespace graphics {
 	}
 
 	void BatchRenderer2D::submit(Sprite* e) {
-		float tid = 0;
+		f32 tid = 0;
 		if(e->hasTexture)
-			tid = (float)submitTexture(e->textureID);
+			tid = (f32)submitTexture(e->textureID);
 		maths::vec4 sprite = e->data;
 
 		buffer->position = *transformationBack * maths::vec3(sprite.x, sprite.y, 1.1f);
@@ -78,57 +78,57 @@ namespace ht { namespace graphics {
 		indexCount += 6;
 	}
 
-	void BatchRenderer2D::submitText(String text, float x, float y, unsigned int color, maths::vec2 scale) {
+	void BatchRenderer2D::submitText(String text, f32 x, f32 y, uint32 color, maths::vec2 scale) {
 		Font& f = FontManager::Get()->getFont();
 
-		unsigned int tid = submitTexture((float)f.atlas->id);
+		uint32 tid = submitTexture((f32)f.atlas->id);
 
 		ftgl::texture_font_t* ftFont = f.font;
 
-		for (unsigned int i = 0; i < text.size - 1; i++) {
+		for (uint32 i = 0; i < text.size - 1; i++) {
 			char c = text[i];
 			ftgl::texture_glyph_t* glyph = texture_font_get_glyph(ftFont, c);
 
 			if (glyph) {
 				if (i > 0) {
-					float kerning = ftgl::texture_glyph_get_kerning(glyph, text[i - 1]);
+					f32 kerning = ftgl::texture_glyph_get_kerning(glyph, text[i - 1]);
 					x += kerning / scale.x;
 				}
-				float x0;
+				f32 x0;
 
 				if(i == 0) x0 = x + (glyph->offset_x / scale.x) / 2.f;
 				else x0 = x + glyph->offset_x / scale.x;
 
-				float y0 = y + glyph->offset_y / scale.y;
-				float x1 = x0 + (glyph->width / scale.x);
-				float y1 = y0 - (glyph->height / scale.y);
+				f32 y0 = y + glyph->offset_y / scale.y;
+				f32 x1 = x0 + (glyph->width / scale.x);
+				f32 y1 = y0 - (glyph->height / scale.y);
 
-				float u0 = glyph->s0;
-				float v0 = glyph->t0;
-				float u1 = glyph->s1;
-				float v1 = glyph->t1;
+				f32 u0 = glyph->s0;
+				f32 v0 = glyph->t0;
+				f32 u1 = glyph->s1;
+				f32 v1 = glyph->t1;
 
 				buffer->position = *transformationBack * maths::vec3(x0, y0, 1.1f);
 				buffer->uv = maths::vec2(u0, v0);
-				buffer->textureID = (float)tid;
+				buffer->textureID = (f32)tid;
 				buffer->color = color;
 				buffer++;
 
 				buffer->position = *transformationBack * maths::vec3(x0, y1, 1.1f);
 				buffer->uv = maths::vec2(u0, v1);
-				buffer->textureID = (float)tid;
+				buffer->textureID = (f32)tid;
 				buffer->color = color;
 				buffer++;
 
 				buffer->position = *transformationBack * maths::vec3(x1, y1, 1.1f);
 				buffer->uv = maths::vec2(u1, v1);
-				buffer->textureID = (float)tid;
+				buffer->textureID = (f32)tid;
 				buffer->color = color;
 				buffer++;
 
 				buffer->position = *transformationBack * maths::vec3(x1, y0, 1.1f);
 				buffer->uv = maths::vec2(u1, v0);
-				buffer->textureID = (float)tid;
+				buffer->textureID = (f32)tid;
 				buffer->color = color;
 				buffer++;
 
@@ -138,13 +138,13 @@ namespace ht { namespace graphics {
 		}
 	}
 
-	void BatchRenderer2D::submitText(String text, float x, float y, vec4 color, vec2 scale) {
-		int r = (int)(color.x * 255.0f);
-		int g = (int)(color.y * 255.0f);
-		int b = (int)(color.z * 255.0f);
-		int a = (int)(color.w * 255.0f);
+	void BatchRenderer2D::submitText(String text, f32 x, f32 y, vec4 color, vec2 scale) {
+		byte r = (byte)(color.x * 255.0f);
+		byte g = (byte)(color.y * 255.0f);
+		byte b = (byte)(color.z * 255.0f);
+		byte a = (byte)(color.w * 255.0f);
 
-		unsigned int col = a << 24 | b << 16 | g << 8 | r;
+		uint32 col = a << 24 | b << 16 | g << 8 | r;
 		submitText(text, x, y, col, scale);
 	}
 
@@ -154,7 +154,7 @@ namespace ht { namespace graphics {
 	}
 
 	void BatchRenderer2D::render() {
-		for (unsigned int i = 0; i < tids.size(); i++) {
+		for (uint32 i = 0; i < tids.size(); i++) {
 			glActiveTexture(GL_TEXTURE0 + i);
 			glBindTexture(GL_TEXTURE_2D, (GLuint)tids[i]);
 		}
@@ -184,16 +184,16 @@ namespace ht { namespace graphics {
 		glEnableVertexAttribArray(SHADER_COLOR_INDEX);
 
 		glVertexAttribPointer(SHADER_VERTEX_INDEX, 3, GL_FLOAT, GL_FALSE, RENDERER_VERTEX_SIZE, (const GLvoid*)0);
-		glVertexAttribPointer(SHADER_UV_INDEX, 2, GL_FLOAT, GL_FALSE, RENDERER_VERTEX_SIZE, (const GLvoid*)(3 * sizeof(GLfloat)));
-		glVertexAttribPointer(SHADER_TID_INDEX, 1, GL_FLOAT, GL_FALSE, RENDERER_VERTEX_SIZE, (const GLvoid*)(5 * sizeof(GLfloat)));
-		glVertexAttribPointer(SHADER_COLOR_INDEX, 4, GL_UNSIGNED_BYTE, GL_TRUE, RENDERER_VERTEX_SIZE, (const GLvoid*)(6 * sizeof(GLfloat)));
+		glVertexAttribPointer(SHADER_UV_INDEX, 2, GL_FLOAT, GL_FALSE, RENDERER_VERTEX_SIZE, (const GLvoid*)(3 * sizeof(f32)));
+		glVertexAttribPointer(SHADER_TID_INDEX, 1, GL_FLOAT, GL_FALSE, RENDERER_VERTEX_SIZE, (const GLvoid*)(5 * sizeof(f32)));
+		glVertexAttribPointer(SHADER_COLOR_INDEX, 4, GL_UNSIGNED_BYTE, GL_TRUE, RENDERER_VERTEX_SIZE, (const GLvoid*)(6 * sizeof(f32)));
 
 
-		GLushort* indices = htnew GLushort[RENDERER_INDICES_SIZE];
+		uint16* indices = htnew uint16[RENDERER_INDICES_SIZE];
 
-		short offset = 0;
+		int16 offset = 0;
 
-		for (int i = 0; i < RENDERER_INDICES_SIZE; i+=6) {
+		for (int32 i = 0; i < RENDERER_INDICES_SIZE; i+=6) {
 			indices[  i  ] = offset + 0;
 			indices[i + 1] = offset + 1;
 			indices[i + 2] = offset + 2;
@@ -206,13 +206,11 @@ namespace ht { namespace graphics {
 
 		ibo = htnew IBO();
 		ibo->bindIBO();
-		glBufferData(GL_ELEMENT_ARRAY_BUFFER, RENDERER_INDICES_SIZE * sizeof(GLushort), indices, GL_STATIC_DRAW);
+		glBufferData(GL_ELEMENT_ARRAY_BUFFER, RENDERER_INDICES_SIZE * sizeof(uint16), indices, GL_STATIC_DRAW);
 
 		ibo->unbindIBO();
 
 		vao.unbindVAO();
 		indexCount = 0;
 	}
-
-
 } }
