+++ Helltooth/src/ht/graphics/window/WindowManager.cpp
@@ -10,18 +10,18 @@ namespace ht { namespace graphics {
 			del window;
 	}
 	
-	unsigned int WindowManager::createWindow(String title, const int width, const int height) {
+	unsigned int WindowManager::createWindow(String title, const uint16 width, const uint16 height) {
 		windows.push_back(htnew Window(title, width, height));
 		return windows.size() - 1;
 	}
 	
-	void WindowManager::resize(GLFWwindow* glfwwindow, int width, int height) {
+	void WindowManager::resize(GLFWwindow* glfwwindow, uint16 width, uint16 height) {
 		for (Window* window : windows)
 			if (window->compare(glfwwindow)) 
 				window->resize(width, height);
 	}
 
-	Window* WindowManager::getWindow(unsigned int id) {
+	Window* WindowManager::getWindow(uint32 id) {
 		if (id > windows.size())
 			HT_FATAL("Window id %i is bigger than vector size %i", id, windows.size());
 		return windows[id];
